# openapi.yaml
openapi: 3.1.0
info:
  title: 'Saturn Platform'
  version: '0.1'
servers:
  -
    url: 'https://app.example.com/api/v1'
    description: 'Saturn Platform Cloud API. Change the host to your own instance if you are self-hosting.'
paths:
...
  '/services/{uuid}/envs':
    get:
      tags:
        - Services
      summary: 'List Envs'
      description: 'List all envs by service UUID.'
      operationId: list-envs-by-service-uuid
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID of the service.'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'All environment variables by service UUID.'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentVariable'
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Services
      summary: 'Create Env'
      description: 'Create env by service UUID.'
      operationId: create-env-by-service-uuid
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID of the service.'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: 'Env created.'
        required: true
        content:
          application/json:
            schema:
              properties:
                key:
                  description: 'The key of the environment variable.'
                  type: string
                value:
                  description: 'The value of the environment variable.'
                  type: string
                is_preview:
                  description: 'The flag to indicate if the environment variable is used in preview deployments.'
                  type: boolean
                is_literal:
                  description: 'The flag to indicate if the environment variable is a literal, nothing espaced.'
                  type: boolean
                is_multiline:
                  description: 'The flag to indicate if the environment variable is multiline.'
                  type: boolean
                is_shown_once:
                  description: "The flag to indicate if the environment variable's value is shown on the UI."
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Environment variable created.'
          content:
            application/json:
              schema:
                properties:
                  uuid: { type: string, example: nc0k04gk8g0cgsk440g0koko }
                type: object
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
      security:
        -
          bearerAuth: []
    patch:
      tags:
        - Services
      summary: 'Update Env'
      description: 'Update env by service UUID.'
      operationId: update-env-by-service-uuid
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID of the service.'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: 'Env updated.'
        required: true
        content:
          application/json:
            schema:
              required:
                - key
                - value
              properties:
                key:
                  description: 'The key of the environment variable.'
                  type: string
                value:
                  description: 'The value of the environment variable.'
                  type: string
                is_preview:
                  description: 'The flag to indicate if the environment variable is used in preview deployments.'
                  type: boolean
                is_literal:
                  description: 'The flag to indicate if the environment variable is a literal, nothing espaced.'
                  type: boolean
                is_multiline:
                  description: 'The flag to indicate if the environment variable is multiline.'
                  type: boolean
                is_shown_once:
                  description: "The flag to indicate if the environment variable's value is shown on the UI."
                  type: boolean
              type: object
      responses:
        '201':
          description: 'Environment variable updated.'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Environment variable updated.' }
                type: object
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
      security:
        -
          bearerAuth: []
  '/services/{uuid}/envs/bulk':
    patch:
      tags:
        - Services
      summary: 'Update Envs (Bulk)'
      description: 'Update multiple envs by service UUID.'
      operationId: update-envs-by-service-uuid
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID of the service.'
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: 'Bulk envs updated.'
        required: true
        content:
          application/json:
            schema:
              required:
                - data
              properties:
                data:
                  type: array
                  items: { properties: { key: { description: 'The key of the environment variable.', type: string }, value: { description: 'The value of the environment variable.', type: string }, is_preview: { description: 'The flag to indicate if the environment variable is used in preview deployments.', type: boolean }, is_literal: { description: 'The flag to indicate if the environment variable is a literal, nothing espaced.', type: boolean }, is_multiline: { description: 'The flag to indicate if the environment variable is multiline.', type: boolean }, is_shown_once: { description: "The flag to indicate if the environment variable's value is shown on the UI.", type: boolean } }, type: object }
              type: object
      responses:
        '201':
          description: 'Environment variables updated.'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Environment variables updated.' }
                type: object
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
      security:
        -
          bearerAuth: []
  '/services/{uuid}/envs/{env_uuid}':
    delete:
      tags:
        - Services
      summary: 'Delete Env'
      description: 'Delete env by UUID.'
      operationId: delete-env-by-service-uuid
      parameters:
        -
          name: uuid
          in: path
          description: 'UUID of the service.'
          required: true
          schema:
            type: string
            format: uuid
        -
          name: env_uuid
          in: path
          description: 'UUID of the environment variable.'
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 'Environment variable deleted.'
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Environment variable deleted.' }
                type: object
        '401':
          $ref: '#/components/responses/401'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
...
    EnvironmentVariable:
      description: 'Environment Variable model'
      properties:
        id:
          type: integer
        uuid:
          type: string
        resourceable_type:
          type: string
        resourceable_id:
          type: integer
        is_literal:
          type: boolean
        is_multiline:
          type: boolean
        is_preview:
          type: boolean
        is_runtime:
          type: boolean
        is_buildtime:
          type: boolean
        is_shared:
          type: boolean
        is_shown_once:
          type: boolean
        key:
          type: string
        value:
          type: string
        real_value:
          type: string
        source_template:
          type: string
          nullable: true
          description: 'Source template file (env_example, env_sample, env_template)'
        version:
          type: string
        created_at:
          type: string
        updated_at:
          type: string
      type: object
...
  responses:
    '400':
      description: 'Invalid token.'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Invalid token.'
            type: object
    '401':
      description: Unauthenticated.
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: Unauthenticated.
            type: object
    '404':
      description: 'Resource not found.'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Resource not found.'
            type: object
    '422':
      description: 'Validation error.'
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
                example: 'Validation error.'
              errors:
                type: object
                example:
                  name: ['The name field is required.']
                  api_url: ['The api url field is required.', 'The api url format is invalid.']
                additionalProperties:
                  type: array
                  items: { type: string }
            type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Go to `Keys & Tokens` / `API tokens` and create a new token. Use the token as the bearer token.'
      scheme: bearer
...